ui/requirements.txt
python-dotenv
Flask==1.0.3
